Update README.rst
I will add changes to the Commands section later. I'll wait to see if you can move that section to another READ.ME and get it working. One suggestion I have is for the Queue Buttons section, we can add a gif to each queue button just to make things more visual but this is more of an optional suggestion.